Fix audit concerns #18-#23
- #18: Drop detail cache TTL from 1800s to 60s
- #19: Add MediaCreateResponse/ClassificationCreateResponse schemas
- #21: Fix price_level column type Numeric → Integer
- #22: Replace colon-delimited cursor with JSON-encoded sort key
- #23: Remove redundant asyncio.ensure_future inside gather
- #17, #20: Deferred (cache validation and region denormalization)

+++ src/dmo/models/database.py
@@ -85,7 +85,7 @@ class Entity(SQLModel, table=True):
     currency: str | None = Field(default=None, sa_column=Column(String(10)))
     price_min: float | None = Field(default=None, sa_column=Column(Numeric(12, 2)))
     price_max: float | None = Field(default=None, sa_column=Column(Numeric(12, 2)))
-    price_level: int | None = Field(default=None, sa_column=Column(Numeric))
+    price_level: int | None = Field(default=None, sa_column=Column(Integer))
 
     is_barrier_free: bool = Field(default=False, sa_column=Column(Boolean, server_default=text("FALSE")))
     wheelchair_accessible: bool | None = Field(default=None, sa_column=Column(Boolean))
